[wptrunner] Split `message-queue.js` from `testharnessreport.js`
This will allow non-testharness tests to use `testdriver.js` without
needing extra scripts. Evaluating `message-queue.js` is idempotent so
that, when using testharness with testdriver, the second inclusion is a
no-op.

Because resource scripts are cached, the size increase should not
meaningfully affect test performance.
