+++ tools/wptrunner/wptrunner/environment.py
@@ -23,8 +23,6 @@ repo_root = os.path.abspath(os.path.join(here, os.pardir, os.pardir, os.pardir))
 sys.path.insert(0, repo_root)
 from tools import localpaths  # noqa: F401
 
-from wptserve.handlers import StringHandler
-
 serve = None
 
 
@@ -226,28 +224,46 @@ class TestEnvironment:
             self.config.aliases,
             self.config)
 
+        testharnessreport_format_args = {
+            "output": self.pause_after_test,
+            "timeout_multiplier": self.testharness_timeout_multipler,
+            "explicit_timeout": "true" if self.debug_info is not None else "false",
+            "debug": "true" if self.debug_test else "false",
+        }
         for path, format_args, content_type, route in [
                 ("testharness_runner.html", {}, "text/html", "/testharness_runner.html"),
                 ("print_pdf_runner.html", {}, "text/html", "/print_pdf_runner.html"),
-                (os.path.join(here, "..", "..", "third_party", "pdf_js", "pdf.js"), None,
+                (os.path.join(here, "..", "..", "third_party", "pdf_js", "pdf.js"), {},
                  "text/javascript", "/_pdf_js/pdf.js"),
-                (os.path.join(here, "..", "..", "third_party", "pdf_js", "pdf.worker.js"), None,
+                (os.path.join(here, "..", "..", "third_party", "pdf_js", "pdf.worker.js"), {},
                  "text/javascript", "/_pdf_js/pdf.worker.js"),
-                (self.options.get("testharnessreport", "testharnessreport.js"),
-                 {"output": self.pause_after_test,
-                  "timeout_multiplier": self.testharness_timeout_multipler,
-                  "explicit_timeout": "true" if self.debug_info is not None else "false",
-                  "debug": "true" if self.debug_test else "false"},
-                 "text/javascript;charset=utf8",
-                 "/resources/testharnessreport.js")]:
-            path = os.path.normpath(os.path.join(here, path))
+                (
+                    self.options.get("testharnessreport", [
+                        # All testharness tests, even those that don't use testdriver, require
+                        # `message-queue.js` to signal completion.
+                        os.path.join("executors", "message-queue.js"),
+                        "testharnessreport.js"]),
+                    testharnessreport_format_args,
+                    "text/javascript;charset=utf8",
+                    "/resources/testharnessreport.js",
+                ),
+                (
+                    [os.path.join(repo_root, "resources", "testdriver.js"),
+                     # Include `message-queue.js` to support testdriver in non-testharness tests.
+                     os.path.join("executors", "message-queue.js"),
+                     "testdriver-extra.js"],
+                    {},
+                    "text/javascript",
+                    "/resources/testdriver.js",
+                ),
+        ]:
+            paths = [path] if isinstance(path, str) else path
+            abs_paths = [os.path.normpath(os.path.join(here, path)) for path in paths]
             # Note that .headers. files don't apply to static routes, so we need to
             # readd any static headers here.
             headers = {"Cache-Control": "max-age=3600"}
-            route_builder.add_static(path, format_args, content_type, route,
+            route_builder.add_static(abs_paths, format_args, content_type, route,
                                      headers=headers)
-
-        route_builder.add_handler("GET", "/resources/testdriver.js", TestdriverLoader())
 
         for url_base, test_root in self.test_paths.items():
             if url_base == "/":
@@ -314,27 +330,6 @@ class TestEnvironment:
                         s.close()
 
         return failed, pending
-
-
-class TestdriverLoader:
-    """A special static handler for serving `/resources/testdriver.js`.
-
-    This handler lazily reads `testdriver{,-extra}.js` so that wptrunner doesn't
-    need to pass the entire file contents to child `wptserve` processes, which
-    can slow `wptserve` startup by several seconds (crbug.com/1479850).
-    """
-    def __init__(self):
-        self._handler = None
-
-    def __call__(self, request, response):
-        if not self._handler:
-            data = b""
-            with open(os.path.join(repo_root, "resources", "testdriver.js"), "rb") as fp:
-                data += fp.read()
-            with open(os.path.join(here, "testdriver-extra.js"), "rb") as fp:
-                data += fp.read()
-            self._handler = StringHandler(data, "text/javascript")
-        return self._handler(request, response)
 
 
 def wait_for_service(logger: StructuredLogger,
